% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\docType{data}
\name{degree}
\alias{degree}
\title{Constant for conversion from degree to radian ####}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
degree
}
\description{
Constant for conversion from degree to radian ####
}
\keyword{datasets}
