% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{DiskInfo}
\alias{DiskInfo}
\title{Return information for tree disk analysed from  TreeRingShape class}
\usage{
DiskInfo(TR., dpi = 1200)
}
\arguments{
\item{TR.}{class of TreeRingShape}

\item{dpi}{Resolution of tree disk image}
}
\value{
data frame of information for tree disk analysed
}
\description{
Return information for tree disk analysed from  TreeRingShape class
}
\seealso{
\code{\link{TreeRingShape}}
}
