% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{dstpp}
\alias{dstpp}
\title{Return vector for distance between adjacent two points}
\usage{
dstpp(x, y)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}
}
\value{
vector for distance between adjacent two points
}
\description{
Return vector for distance between adjacent two points
}
\examples{
l.<-TR@L[[1]]
plot(l.)
x<-l.[,1] ;y<-l.[,2]
dstpp(x,y)
}
