% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Ldeg360}
\alias{Ldeg360}
\title{Return a vector of center angle 0 to 360(degree) for x y coordinate vector}
\usage{
Ldeg360(x, y)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}
}
\value{
a vector of center angle 0 to 360(degree) for x y coordinate vector
}
\description{
Return a vector of center angle 0 to 360(degree) for x y coordinate vector
}
\examples{
 xy <-TR@L[[1]]
 plot(Ldeg360(xy[,1],xy[2]))

}
