% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Llist2dataframe}
\alias{Llist2dataframe}
\title{Convert from a list of tree rings polygons (L) to data frame
to a data frame with no.,year,x,y,r(radius),radian(center angle),degree.
The data frame is sorted by degree(0 to 360).}
\usage{
Llist2dataframe(L)
}
\arguments{
\item{L}{list of tree ring lines}
}
\value{
data frame
}
\description{
Convert from a list of tree rings polygons (L) to data frame
to a data frame with no.,year,x,y,r(radius),radian(center angle),degree.
The data frame is sorted by degree(0 to 360).
}
\examples{
L_ <- Llist2dataframe(TR@L)
head(L_) ; tail(L_)
}
