% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lmove}
\alias{Lmove}
\title{Move the tree rings coordinates based on P00 (x,y movement coordinates).}
\usage{
Lmove(L, P00 = P00)
}
\arguments{
\item{L}{a list of tree rings(x,y coordinates).}

\item{P00}{x, y coordinates of a center point (usually a pith).}
}
\value{
moved L to center point 0,0
}
\description{
Move the tree rings coordinates based on P00 (x,y movement coordinates).
}
\examples{
Lplot(TR@L)
sapply(Lmove(TR@L,c(3000,-3000)),lines,col="blue")
}
