% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lplot}
\alias{Lplot}
\title{Plot a graphics of tree rings}
\usage{
Lplot(L, rn = 1:length(L), col = "red", ...)
}
\arguments{
\item{L}{a list of tree rings polygon coordinates (X,Y)}

\item{rn}{vector of ring number of list (L), default 1:length(L)}

\item{col}{color of plot}

\item{...}{other parameters to be passed through to plotting functions}
}
\description{
Plot a graphics of tree rings
}
\examples{

Lplot(TR@L2,main=TR@L2_filename)
Lplot(TR@L2,rn=1:50,col='blue',main=TR@L2_filename)
}
