% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lplot2}
\alias{Lplot2}
\title{Draw a graphics of tree rings by 1 ring (3*3 in a screen)}
\usage{
Lplot2(L, i.ring = 1:length(L), nrow = 3, ncol = 3, ask = "FALSE", ...)
}
\arguments{
\item{L}{a list of tree rings polygon coordinates (X,Y)}

\item{i.ring}{integer vector, tree ring number for drawing}

\item{nrow}{par(mfrow=c(nrow,ncol))}

\item{ncol}{par(mfrow=c(nrow,ncol))}

\item{ask}{logical; if TRUE, the user is asked before each plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
No return value, only draw tree ring plot.
}
\description{
Draw a graphics of tree rings by 1 ring (3*3 in a screen)
}
\examples{
Lplot2(TR@L,i.ring=1:9, nrow=1,ncol=1,type='b')
Lplot2(TR@L,type='b')
}
