% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lrad.plot}
\alias{Lrad.plot}
\title{Check center angle of points to input order}
\usage{
Lrad.plot(L, i.ring = 1:9)
}
\arguments{
\item{L}{list of tree rings}

\item{i.ring}{integer vector, tree ring number for drawing}
}
\description{
Check center angle of points to input order
}
\examples{
slotNames(TR)
Lplot(TR@L)
str(TR@L)
Lrad.plot(TR@L,11:19)

}
