% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lrn}
\alias{Lrn}
\title{Return a ring number of tree ring polygons list (L) from year}
\usage{
Lrn(L, yr)
}
\arguments{
\item{L}{tree ring polygons list (L)}

\item{yr}{years (or rings)}
}
\value{
a ring number of tree ring polygons list (L)
}
\description{
Return a ring number of tree ring polygons list (L) from year
}
\examples{
Lrn(TR@L,168)  # 168 is the formation year (from outermost) of the tree ring

}
