% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lsort}
\alias{Lsort}
\title{Sort  x,y coordinates of a tree ring line with center angle of each point}
\usage{
Lsort(l.)
}
\arguments{
\item{l.}{x,y coordinates matrix (ncol=2) or data.frame of an tree ring.}
}
\value{
ordered with center angle of each point
}
\description{
Sort  x,y coordinates of a tree ring line with center angle of each point
}
\examples{
i<-seq(0,2*pi,0.1)
l.<-data.frame(x=sin(i),y=cos(i))
l.[10,]<-l.[20,]
plot(l.,type="b")
plot(Lsort(l.),type="b")

}
