% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lsort_all}
\alias{Lsort_all}
\title{Sort  x,y coordinates of tree ring lines with center angle of each point
apply Lsort to list of tree ring lines}
\usage{
Lsort_all(L)
}
\arguments{
\item{L}{a list of tree ring lines (x,y)}
}
\value{
a list of tree ring lines (x,y) ordered with center angle of each point
}
\description{
Sort  x,y coordinates of tree ring lines with center angle of each point
apply Lsort to list of tree ring lines
}
\examples{
str(Lsort_all(TR@L))

}
