% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{new_classTreeRingShape}
\alias{new_classTreeRingShape}
\title{Initial setting of  a new classTreeRingShape (TR)}
\usage{
new_classTreeRingShape(
  P_filename,
  L_filename,
  L2_filename,
  P_id.tag = "id",
  P_ring.tag = "ring",
  L_ring.tag = "ring"
)
}
\arguments{
\item{P_filename}{file name of shape file (P) for tree ring points}

\item{L_filename}{file name of shape file (L) for tree ring lines}

\item{L2_filename}{file name of shape file (L2) for tree ring lines interpolated}

\item{P_id.tag}{column name of id in shape file (P), default is 'id'}

\item{P_ring.tag}{column name of ring no.(ordinaly year,outermost=0) in shape file (L), default is 'ring'}

\item{L_ring.tag}{column name of ring no.(ordinaly year,outermost=0) in shape file (L), default is 'ring'}
}
\value{
generated new object from classTreeRingShape
}
\description{
Initial setting of  a new classTreeRingShape (TR)
}
\examples{
TR_<-new_classTreeRingShape(
P_filename='Abies_277_h400_TreeRing_Points.shp',
L_filename='Abies_277_h400_TreeRing_Representative.shp',
L2_filename='Abies_277_h400_TreeRing.shp',
P_id.tag='id',
P_ring.tag='ring',
L_ring.tag='ring')

 TR_
 slotNames(TR_)
 str(TR_)



}
