% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_TreeRing}
\alias{plot_TreeRing}
\title{Draw a plot of tree rings
This function draws Tree rings of a disk from x, y list(x,y) with name of year.}
\usage{
plot_TreeRing(L, year = 0, ...)
}
\arguments{
\item{L}{list(x,y) of Tree ring coordinates with name of year}

\item{year}{name of  column of Tree ring year (0(cambium),1,2,....,n(pith))}

\item{...}{other parameters to be passed through to plotting functions}
}
\value{
No return value, only draw tree ring plot.
}
\description{
Draw a plot of tree rings
This function draws Tree rings of a disk from x, y list(x,y) with name of year.
}
\examples{
names(TR@L)
plot_TreeRing(TR@L)
plot_TreeRing(TR@L,year=10,type='l',col='blue')
}
