% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_TreeRing_df}
\alias{plot_TreeRing_df}
\title{plot_TreeRing_df
Draw a Tree ring of a disk from data frame(x,y,year)}
\usage{
plot_TreeRing_df(df, year = 0, year_label = "yr")
}
\arguments{
\item{df}{name of a data frame}

\item{year}{integer vector of years to draw tree rings}

\item{year_label}{name of  column of Tree ring year (0(cambium),1,2,....,n(pith))}
}
\value{
No return value, only draw tree ring plot.
}
\description{
plot_TreeRing_df
Draw a Tree ring of a disk from data frame(x,y,year)
}
\examples{
TR@L_ <- Llist2dataframe(TR@L)     # data frame of tree rings
plot_TreeRing_df(TR@L_, year =1)

}
