% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_TreeRings_df}
\alias{plot_TreeRings_df}
\title{Plot tree rings from data fame
This function draws Tree rings of a disk from data frame(x,y,year).}
\usage{
plot_TreeRings_df(df, year_label = "yr")
}
\arguments{
\item{df}{name of a data frame}

\item{year_label}{name of  column of Tree ring year (0(cambium),1,2,....,n(pith))}
}
\description{
Plot tree rings from data fame
This function draws Tree rings of a disk from data frame(x,y,year).
}
\examples{
df <- TR@L_  # data frame of tree rings
names(df)
plot_TreeRings_df(df)

}
\seealso{
\code{\link{Llist2dataframe}}  for the data frame
}
