% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_year_RingArea}
\alias{plot_year_RingArea}
\title{Plot and return data frame of  year_disk area and year_Tree ring area}
\usage{
plot_year_RingArea(L2, yr_end = 2018)
}
\arguments{
\item{L2}{list of tree rings}

\item{yr_end}{outermost year of tree ring}
}
\value{
list of Year_DiskArea and Year_TreeRingArea
}
\description{
Plot and return data frame of  year_disk area and year_Tree ring area
}
\seealso{
\code{\link{TreeRingsInterpolation}}
}
