% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_TreeRingPoints}
\alias{ReadShapefile_TreeRingPoints}
\title{Read a shape file of Tree Ring Points ( P : radial input and correction points)}
\usage{
ReadShapefile_TreeRingPoints(
  filename = "points277_h600",
  id.tag = "id",
  ring.tag = "ring"
)
}
\arguments{
\item{filename}{is a file name of Tree ring points (shape file )
The extension (.shp) is unnecessary.}

\item{id.tag}{string, column name of id (attribute table)}

\item{ring.tag}{string, column name of ring years  (0 is cambium layer)}
}
\value{
a data frame of TreeRingPoints (radial input and correction points)
}
\description{
Read a shape file of Tree Ring Points ( P : radial input and correction points)
}
\examples{
\dontrun{

# sample data of 'Abies_277_h400' can be download from
#https://www.sanchikanri.com/treering/Abies_277_h400.zip

file.path <- '../Abies_277_h400/Abies_277_h400_TreeRing_Points'
ReadShapefile_TreeRingPoints(file.path,id.tag='id',ring.tag='ring')

}

}
