% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_TreeRings}
\alias{ReadShapefile_TreeRings}
\title{Read Shapefile_TreeRings}
\usage{
ReadShapefile_TreeRings(
  filename = "Abies_277_h400_TreeRing_Representative",
  ring.tag = "ring"
)
}
\arguments{
\item{filename}{is a file name of shape file written to disk.
The extension (.shp) is unnecessary.}

\item{ring.tag}{string, column name of ring years  (0 is cambium layer)}
}
\value{
a data frame of TreeRingPoints (radial input and correction points)
}
\description{
Read Shapefile_TreeRings
}
\examples{
\dontrun{
file.path <- '../Abies_277_h400/Abies_277_h400_TreeRing_Representative'
Lplot(ReadShapefile_TreeRings(file.path))
}

}
