% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{data}
\name{TR}
\alias{TR}
\title{A sample object of class TreeRingShape}
\format{
An object of class \code{classTreeRingShape} of length 1.
}
\usage{
TR
}
\description{
The data set contains tree ring shape data for Abies_277_h400 sampled from Tateyama, central Japan.
Its disk image and shape files can be download from  https://www.sanchikanri.com/treering/Abies_277_h400.zip
It's intended to demonstrate the structure and use of `TreeRingShape` class objects within the package.
}
\examples{
# Access basic information about the TreeRingShape object
slotNames(TR)
str(TR)
# Plot the tree ring shape data
Lplot(TR@L)
}
\keyword{datasets}
