% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{TreeRingShape}
\alias{TreeRingShape}
\title{Generates a object (TR) of classTreeRingShape}
\usage{
TreeRingShape(
  P_filename,
  L_filename,
  L2_filename,
  P_id.tag = "id",
  P_ring.tag = "ring",
  L_ring.tag = "ring"
)
}
\arguments{
\item{P_filename}{file name of shape file (P) for tree ring points  (without extention)}

\item{L_filename}{file name of shape file (L) for tree ring lines  (without extention)}

\item{L2_filename}{file name of shape file (L2) for tree ring lines interpolated (without extention)}

\item{P_id.tag}{column name of id in shape file (P), default is 'id'}

\item{P_ring.tag}{column name of ring no.(ordinaly year,outermost=0) in shape file (L), default is 'ring'}

\item{L_ring.tag}{column name of ring no.(ordinaly year,outermost=0) in shape file (L), default is 'ring'}
}
\value{
generated new object from classTreeRingShape
}
\description{
Generates a object (TR) of classTreeRingShape
}
\examples{
\dontrun{
(wd. <- getwd())
setwd('../Abies_277_h400')
dir()
TR.<-TreeRingShape(
P_filename='Abies_277_h400_TreeRing_Points',
L_filename='Abies_277_h400_TreeRing_Representative',
L2_filename='Abies_277_h400_TreeRing',
P_id.tag='id',P_ring.tag='ring',
L_ring.tag='ring')

 TR.
 slotNames(TR.)
 str(TR.)
 dev.new()
Lplot(TR.@L2)

}


}
