% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsLines}
\alias{TreeRingsLines}
\title{Read representative tree ring lines from shape files}
\usage{
TreeRingsLines(TR)
}
\arguments{
\item{TR}{is tree ring  class (classTreeRingShape )}
}
\value{
TR (TreeRing class TR@L<-L ; TR@L_<-L_ ; TR@YR_L <-YR_L ; TR@ln <- ln)
}
\description{
Read representative tree ring lines from shape files
}
\examples{
 \dontrun{
 setwd('../Abies_277_h400')
 TreeRingsLines(TR)
 }

}
