% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsPoints}
\alias{TreeRingsPoints}
\title{Read TreeRingsPoints shape file, check and save parameters}
\usage{
TreeRingsPoints(TR)
}
\arguments{
\item{TR}{is tree ring  class (classTreeRingShape )

#@return list of  (P,P00,YR_P,n_id,YR_P,n_YR)}
}
\description{
Read TreeRingsPoints shape file, check and save parameters
}
\examples{
\dontrun{
P_filename<-'points277_h600'\t\t#####shape file name (without extention 'abc.shp' -> 'abc')
TreeRingsPoints(TR)
#' }
}
