% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{WriteShapefile_TreeRings}
\alias{WriteShapefile_TreeRings}
\title{Write a shapefile of interpolated tree rings}
\usage{
WriteShapefile_TreeRings(L2, filename = "test")
}
\arguments{
\item{L2}{is as list of Tree ring polygons (X, Y)}

\item{filename}{is a file name of shape file written to disk.
The extension (.shp) is unnecessary.}
}
\value{
data of Shapefile
}
\description{
Write a shapefile of interpolated tree rings
}
\examples{

\dontrun{
file.path <- '../Abies_277_h400_test'
WriteShapefile_TreeRings (TR@L, file.path)
}
}
