% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_TreeRings}
\alias{ReadShapefile_TreeRings}
\title{Read Shapefile_TreeRings}
\usage{
ReadShapefile_TreeRings(
  filename = "Abies_277_h400_TreeRing_Representative.shp",
  ring.tag = "ring"
)
}
\arguments{
\item{filename}{is a file name(path) of shape file written to disk.}

\item{ring.tag}{string, column name of ring years  (0 is cambium layer)}
}
\value{
a list of tree ring lines
}
\description{
Read Shapefile_TreeRings
}
\examples{
\dontrun{
filename <- '../Abies_277_h400/Abies_277_h400_TreeRing_Representative.shp'
Lplot(ReadShapefile_TreeRings(filename))
}

}
