% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_TreeRings}
\alias{ReadShapefile_TreeRings}
\title{Read Shapefile_TreeRings}
\usage{
ReadShapefile_TreeRings(
  filename = "Abies_277_h400_TreeRing_Representative.shp",
  ring.tag = "ring"
)
}
\arguments{
\item{filename}{a file name(path) of shape file written to disk.}

\item{ring.tag}{string, column name of ring years  (0 is cambium layer)}
}
\value{
a list of tree ring lines
}
\description{
Read Shapefile_TreeRings
}
\examples{
.dir <- system.file("shp",package = "TreeRingShape")
.file <- "Abies_277_h400_TreeRing_Representative.shp"
filename <- paste(.dir,.file,sep="/")
sf.L<-sf::st_read(filename)
plot(sf.L)
Lplot(ReadShapefile_TreeRings(filename))

}
