% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{TreeRingIndex}
\alias{TreeRingIndex}
\title{Calculate tree ring index from chronosequence data (year,growth)}
\usage{
TreeRingIndex(ya, spar = 0.8)
}
\arguments{
\item{ya}{data frame of chronosequence data (year,growth)}

\item{spar}{smoothing parameter of spline curve}
}
\value{
list spline ; fitting parameter of Spline curve  ,
idx ; data.frame(year,TreeRingIndex)
}
\description{
Calculate tree ring index from chronosequence data (year,growth)
}
\references{
Cook, E., & Peters, K. (1981). The smoothing spline,
a new approach to standardising forest interior tree-ring.Trre-ring Bulletin,
41, 45–53.
}
\seealso{
\code{\link{TreeRingsInterpolation}}
}
