% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsPoints}
\alias{TreeRingsPoints}
\title{Read TreeRingsPoints shape file, check and save parameters}
\usage{
TreeRingsPoints(TR)
}
\arguments{
\item{TR}{a tree ring  class (classTreeRingShape )

#@return a list of  (P,P00,YR_P,n_id,YR_P,n_YR)}
}
\description{
Read TreeRingsPoints shape file, check and save parameters
}
\examples{
\dontrun{
 setwd('../Abies_277_h400') #set working directory with shape files
 TreeRingsPoints(TR)
#' }
}
