% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsPoints}
\alias{TreeRingsPoints}
\title{Read TreeRingsPoints shape file, check and save parameters}
\usage{
TreeRingsPoints(TR)
}
\arguments{
\item{TR}{a tree ring  class (classTreeRingShape )}
}
\value{
a list of  (P,P00,YR_P,n_id,YR_P,n_YR)
}
\description{
Read TreeRingsPoints shape file, check and save parameters
}
\examples{
 #  didectory of tree ring shapefiles
 .dir <- system.file("shp",package = "TreeRingShape")

# path of  P_filename
.file <- "Abies_277_h400_TreeRing_Points.shp"
 TR_@P_filename <- paste(.dir,.file,sep="/")

  TreeRingsPoints(TR_)@P

}
