% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{WriteShapefile_TreeRings}
\alias{WriteShapefile_TreeRings}
\title{Write a shapefile of interpolated tree rings}
\usage{
WriteShapefile_TreeRings(L2, filename = "test.shp")
}
\arguments{
\item{L2}{is as list of Tree ring polygons (X, Y)}

\item{filename}{is a shape file(path) name written to disk.}
}
\value{
No return value, called for side effects.
}
\description{
Write a shapefile of interpolated tree rings
}
\examples{

\donttest{#'
WriteShapefile_TreeRings (TR@L, tempfile("TreeRingShape_test",fileext = ".shp"))
dir(tempdir())
}
}
